+++ frontend/src/pages/team/poule/TeamWinrates.tsx
@@ -15,7 +15,7 @@ export default function TeamWinRates({ poule }: { poule: DetailedPouleInfo }) {
     <Paper elevation={4}>
       <Typography variant="h4">Winstpercentages</Typography>
       <hr />
-      <ButtonGroup className="select-winrate-buttons">
+      <ButtonGroup className="select-metric-button-group">
         <Button variant={selectedMetric === 'matchWinRate' ? 'contained' : 'outlined'} onClick={() => setSelectedMetric('matchWinRate')}>Wedstrijden</Button>
         <Button variant={selectedMetric === 'setWinRate' ? 'contained' : 'outlined'} onClick={() => setSelectedMetric('setWinRate')}>Sets</Button>
         <Button variant={selectedMetric === 'pointWinRate' ? 'contained' : 'outlined'} onClick={() => setSelectedMetric('pointWinRate')}>Punten</Button>